containers test: Exercise StopOnDisconnect and StopOnNotify
In the STOP_SERVER_VIA_NOTIFY test, the server would stop whenever
either the mock container manager disconnects from the bus or the
notify fd is closed, whichever comes first. In the
STOP_SERVER_VIA_NOTIFY_NOT_MANAGER test, we assert that the server does
not stop when the mock container manager disconnects, only when the
notify fd is closed.

Signed-off-by: Simon McVittie <smcv@collabora.com>